Fix leak for symbolic link in exfat_lookup
While exfat_lookup() for symbolic file, we Should not alloc memory
to EXFAT_I(inode)->target since the corredspoding exfat inode info
is still is in memory and EXFAT_I(inode)->target has not released
yet. If we do so, memory leak would happen. Therefore we only alloc
it if not null.

Change-Id: I36c9355cdc7ac9a27ebd5337813416cc3c62af9a
Reviewed-by: Ethan Wu <ethanwu@synology.com>
Signed-off-by: Chung-Chiang Cheng <cccheng@synology.com>